#324: added state machine for ssh

+++ agent/src/runner/ssh.ts
@@ -28,14 +28,21 @@ export interface SSHConfig {
 }
 
 export class SSH extends EventEmitter {
-  private _serverActive = false
-  private _clientActive = false
   private _log: AgentRunnerLogger
   private _sshProcessingChanges?: boolean
   private _pendingConfig: SSHConfig | null = null
   private _sshClientManager: ProcessManager
   private _sshServerManager: ProcessManager
   private _privateKeyPath: string
+
+  private readonly STATUS_IDLE = 'idle'
+  private readonly STATUS_STARTING = 'starting'
+  private readonly STATUS_RUNNING = 'running'
+  private readonly STATUS_STOPPING = 'stopping'
+
+  private _serverStatus: string
+  private _clientStatus: string
+  private _active: boolean
 
   public constructor(log: AgentRunnerLogger) {
     super()
@@ -45,6 +52,10 @@ export class SSH extends EventEmitter {
     this._sshServerManager = new ProcessManager('ssh-server', this._log)
     this._sshServerManager.maxRetryCount = 3
     this._privateKeyPath = path.resolve(__dirname, '../../keys/tmp_private_key')
+
+    this._serverStatus = this.STATUS_IDLE
+    this._clientStatus = this.STATUS_IDLE
+    this._active = false
   }
 
   private _debug(...args: any[]): void {
@@ -64,7 +75,47 @@ export class SSH extends EventEmitter {
   }
 
   public statusUpdate(): void {
-    this.emit('statusChanged', this._clientActive && this._serverActive)
+    this.emit('statusChanged', this._active)
+  }
+
+  public _statusChanged(): void {
+    const active = this._clientStatus === this.STATUS_RUNNING
+        && this._serverStatus === this.STATUS_RUNNING
+    if (active !== this._active) {
+      this._info(`ssh active: ${this._active} ==> ${active}`)
+      this._active = active
+      this.statusUpdate()
+    }
+  }
+
+  private _clientStatusChanged(status: string): void {
+    if (status === this._clientStatus)
+      return
+    this._info(`ssh-client status changed: ${this._clientStatus} ==> ${status}`)
+    if (status === this.STATUS_IDLE) {
+      if (this._clientStatus === this.STATUS_STARTING
+          || this._clientStatus === this.STATUS_RUNNING) {
+        this._debug(`ssh-client exited unexpected, shutting down ssh-server`)
+        this.stopServer()
+      }
+    }
+    this._clientStatus = status
+    this._statusChanged()
+  }
+
+  private _serverStatusChanged(status: string): void {
+    if (status === this._serverStatus)
+      return
+    this._info(`ssh-server status changed: ${this._serverStatus} ==> ${status}`)
+    if (status === this.STATUS_IDLE) {
+      if (this._serverStatus === this.STATUS_STARTING
+          || this._serverStatus === this.STATUS_RUNNING) {
+        this._debug(`ssh-server exited unexpected, shutting down ssh-client`)
+        this.stopClient()
+      }
+    }
+    this._serverStatus = status
+    this._statusChanged()
   }
 
   private _exec(cmd: string): Promise<void> {
@@ -79,73 +130,82 @@ export class SSH extends EventEmitter {
   }
 
   public async startServer(options: SSHServerOptions): Promise<void> {
-    if (!this._serverActive) {
-      const getentResult = execReturnStdout(`getent passwd ${options.user}`)
-      if (!getentResult) {
-        throw new Error(`Failed to get home directory of user ${options.user}`)
-      }
-      const userInfo = getUserInfo(options.user)
-      const userHome = getentResult.split(':')[5]
-      const userSSHPath = `${userHome}/.ssh`
-      try {
-        if (!fs.existsSync(userSSHPath)) {
-          fs.mkdirSync(userSSHPath)
-          fs.chownSync(userSSHPath, userInfo.uid, userInfo.gid)
-          fs.chmodSync(userSSHPath, 0o700)
-        }
-        const authorizedKeys = `${userSSHPath}/authorized_keys`
-        if (!fs.existsSync(authorizedKeys)) {
-          fs.writeFileSync(authorizedKeys, options.publicKey, 'utf8')
-          fs.chownSync(authorizedKeys, userInfo.uid, userInfo.gid)
-          fs.chmodSync(authorizedKeys, 0o600)
-        } else {
-          const keys = fs.readFileSync(authorizedKeys, 'utf8')
-          if (keys.indexOf(options.publicKey) === -1) {
-            fs.appendFileSync(authorizedKeys, options.publicKey, 'utf8')
-          }
-        }
-      } catch (err) {
-        throw new Error(`Failed to save public key: ${err.message}`)
-      }
-
-      this._sshServerManager.on('started', () => {
-        this._serverActive = true
-        this._debug(`ssh-server active: ${this._serverActive}`)
-        this.statusUpdate()
-      })
-
-      this._sshServerManager.on('permanentlyTerminated', () => {
-        this._serverActive = false
-        this._debug(`ssh-server active: ${this._serverActive}`)
-        this.statusUpdate()
-      })
-      this._sshServerManager.startedIfTraceContains(
-        `Server listening on :: port ${options.port}`,
-        30 * 1000
-      )
-
-      const args = [
-        '-D',
-        '-e',
-        '-o LogLevel=DEBUG',
-        `-o AllowUsers=${options.user}`,
-        '-o PasswordAuthentication=no',
-        '-o ClientAliveInterval=30',
-        '-o ClientAliveCountMax=3',
-        `-p ${options.port}`
-      ]
-
-      return this._sshServerManager.start('/usr/sbin/sshd', args)
-    } else {
-      this._info('ssh-server already started')
+    if (this._serverStatus !== this.STATUS_IDLE) {
+      this._debug(`Cannot start ssh-server, unexpected status: ${this._serverStatus}`)
+      return
     }
+
+    const getentResult = execReturnStdout(`getent passwd ${options.user}`)
+    if (!getentResult) {
+      throw new Error(`Failed to get home directory of user ${options.user}`)
+    }
+    const userInfo = getUserInfo(options.user)
+    const userHome = getentResult.split(':')[5]
+    const userSSHPath = `${userHome}/.ssh`
+    try {
+      if (!fs.existsSync(userSSHPath)) {
+        fs.mkdirSync(userSSHPath)
+        fs.chownSync(userSSHPath, userInfo.uid, userInfo.gid)
+        fs.chmodSync(userSSHPath, 0o700)
+      }
+      const authorizedKeys = `${userSSHPath}/authorized_keys`
+      if (!fs.existsSync(authorizedKeys)) {
+        fs.writeFileSync(authorizedKeys, options.publicKey, 'utf8')
+        fs.chownSync(authorizedKeys, userInfo.uid, userInfo.gid)
+        fs.chmodSync(authorizedKeys, 0o600)
+      } else {
+        const keys = fs.readFileSync(authorizedKeys, 'utf8')
+        if (keys.indexOf(options.publicKey) === -1) {
+          fs.appendFileSync(authorizedKeys, options.publicKey, 'utf8')
+        }
+      }
+    } catch (err) {
+      throw new Error(`Failed to save public key: ${err.message}`)
+    }
+
+    this._sshServerManager.on('started', () => {
+      this._serverStatusChanged(this.STATUS_RUNNING)
+    })
+
+    this._sshServerManager.on('permanentlyTerminated', () => {
+      this._serverStatusChanged(this.STATUS_IDLE)
+    })
+    this._sshServerManager.startedIfTraceContains(
+      `Server listening on :: port ${options.port}`,
+      30 * 1000
+    )
+
+    const args = [
+      '-D',
+      '-e',
+      '-o LogLevel=DEBUG',
+      `-o AllowUsers=${options.user}`,
+      '-o PasswordAuthentication=no',
+      '-o ClientAliveInterval=30',
+      '-o ClientAliveCountMax=3',
+      `-p ${options.port}`
+    ]
+
+    this._serverStatusChanged(this.STATUS_STARTING)
+    return this._sshServerManager.start('/usr/sbin/sshd', args)
   }
 
   public async stopServer(): Promise<void> {
+    if (this._serverStatus === this.STATUS_STOPPING ||
+        this._serverStatus === this.STATUS_IDLE) {
+      this._debug(`Cannot stop ssh-server, unexpected status: ${this._serverStatus}`)
+      return
+    }
+    this._serverStatusChanged(this.STATUS_STOPPING)
     return this._sshServerManager.stop()
   }
 
   public async startClient(options: SSHClientOptions): Promise<void> {
+    if (this._clientStatus !== this.STATUS_IDLE) {
+      this._debug(`Cannot start ssh-client, unexpected status: ${this._clientStatus}`)
+      return
+    }
+
     const userInfo = getUserInfo(options.user)
     const privateKeyPath = this._privateKeyPath
     fs.writeFileSync(privateKeyPath, options.privateKey, 'utf8')
@@ -165,25 +225,29 @@ export class SSH extends EventEmitter {
     ]
 
     this._sshClientManager.on('started', () => {
-      this._clientActive = true
-      this._debug(`ssh-client active: ${this._clientActive}`)
-      this.statusUpdate()
+      this._clientStatusChanged(this.STATUS_RUNNING)
     })
 
     this._sshClientManager.on('permanentlyTerminated', () => {
-      this._clientActive = false
-      this._debug(`ssh-client active: ${this._clientActive}`)
-      this.statusUpdate()
+      this._clientStatusChanged(this.STATUS_IDLE)
     })
     this._sshClientManager.startedIfTraceContains(
       'All remote forwarding requests processed',
       // It may take up to 2 mins to timeout in connecting
       3 * 60 * 1000
     )
+
+    this._clientStatusChanged(this.STATUS_STARTING)
     return this._sshClientManager.start('/usr/bin/ssh', args, userInfo)
   }
 
   public async stopClient(): Promise<void> {
+    if (this._clientStatus === this.STATUS_STOPPING ||
+        this._clientStatus === this.STATUS_IDLE) {
+      this._debug(`Cannot stop ssh-client, unexpected status: ${this._clientStatus}`)
+      return
+    }
+    this._clientStatusChanged(this.STATUS_STOPPING)
     await this._sshClientManager.stop()
     if (fs.existsSync(this._privateKeyPath)) {
       fs.unlinkSync(this._privateKeyPath)
@@ -231,10 +295,6 @@ export class SSH extends EventEmitter {
         }
       } catch (err) {
         this._error(`process ssh changes failed: ${err.message}`)
-        if (pendingConfig.enable) {
-          await this.stopClient()
-          await this.stopServer()
-        }
       }
     }
     this._sshProcessingChanges = false
